aider: Added scan and connect functionality to `ScaleTableViewController.swift`.
# Context:
USER: great now can you add the scan and connect functionality?
ASSISTANT: Sure, we will need to modify the `ScaleTableViewController.swift` file to add the scan and connect functionality.

First, we will implement the `startScanning` method to start scanning for scales using the `AcaiaManager`'s `startScan` method. We will also add an observer for the `AcaiaScaleListChangedNotification` notification to update the list of scales when it changes.

Next, we will implement the `tableView(_:didSelectRowAt:)` method to connect to the selected scale when a row is selected.

```swift
AiderTest/AiderTest/ScaleTableViewController.swift
<<<<<<< ORIGINAL
    func startScanning() {
        // Start scanning for scales
    }

    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
        let scale = scales[indexPath.row]
        // Connect to the selected scale
    }
=======
    func startScanning() {
        AcaiaManager.shared().startScan(10)
        NotificationCenter.default.addObserver(self, selector: #selector(scaleListChanged), name: NSNotification.Name.AcaiaScaleListChangedNotification, object: nil)
    }

    @objc func scaleListChanged() {
        scales = AcaiaManager.shared().scaleList as? [AcaiaScale] ?? []
        tableView.reloadData()
    }

    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
        let scale = scales[indexPath.row]
        scale.connect()
    }
>>>>>>> UPDATED
```

Please note that this code assumes that the `AcaiaManager`'s `scaleList` property is an array of `AcaiaScale` objects. If this is not the case, the code may need to be adjusted.