aider: Registered a basic `UITableViewCell` with the identifier "ScaleCell" in `ScaleTableViewController.swift`.
# Context:
USER: Thread 1: "unable to dequeue a cell with identifier ScaleCell - must register a nib or a class for the identifier or connect a prototype cell in a storyboard"
ASSISTANT: This error occurs because the table view is trying to dequeue a cell with the identifier "ScaleCell", but it can't find any. You need to register a cell with this identifier before you can dequeue it.

Here's how you can fix this:

1. Modify `ScaleTableViewController.swift` to register a cell with the identifier "ScaleCell".

```swift
AiderTest/AiderTest/ScaleTableViewController.swift
<<<<<<< ORIGINAL
    override func viewDidLoad() {
        super.viewDidLoad()

        _setRefreshControl()
        _addAcaiaEventsObserver()
    }
=======
    override func viewDidLoad() {
        super.viewDidLoad()

        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "ScaleCell")

        _setRefreshControl()
        _addAcaiaEventsObserver()
    }
>>>>>>> UPDATED
```

This change registers a basic `UITableViewCell` with the identifier "ScaleCell". If you have a custom cell that you want to use, replace `UITableViewCell.self` with your custom cell class.