aider: Set start button constraints to bottom left corner of the screen.
# Context:
USER: great, our start button for starting out chart in a bit inacessible in the top left corner of our screen. can we constrain it to the bottom left 20points from the left side of the screen and 20 points above the safe area. I'd like this to be in the bottom left no matter the rotation of the phone.
ASSISTANT: To implement this feature, we need to modify the `AiderTest/AiderTest/ChartViewController.swift` file.

We need to update the `setupUI` method to set the constraints of the start button to the bottom left of the screen, 20 points from the left side of the screen and 20 points above the safe area.

Here are the changes:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func setupUI() {
        // Create and setup the start button
        startButton = UIButton(type: .system)
        startButton.setTitle("Start", for: .normal)
        startButton.addTarget(self, action: #selector(startButtonTapped), for: .touchUpInside)
        startButton.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(startButton)

        // Create and setup the weight chart
        weightChart = LineChartView()
        weightChart.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(weightChart)

        // Create and setup the flow chart
        flowChart = LineChartView()
        flowChart.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(flowChart)
    }
=======
    func setupUI() {
        // Create and setup the start button
        startButton = UIButton(type: .system)
        startButton.setTitle("Start", for: .normal)
        startButton.addTarget(self, action: #selector(startButtonTapped), for: .touchUpInside)
        startButton.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(startButton)

        // Create and setup the weight chart
        weightChart = LineChartView()
        weightChart.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(weightChart)

        // Create and setup the flow chart
        flowChart = LineChartView()
        flowChart.translatesAutoresizingMaskIntoConstraints = false
        view.addSubview(flowChart)

        // Setup constraints
        NSLayoutConstraint.activate([
            startButton.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
            startButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -20)
        ])
    }
>>>>>>> UPDATED
```

This will ensure that the start button is always in the bottom left no matter the rotation of the phone.