aider: Modified the `logData` method to update the charts on the main thread.
# Context:
USER: great, the button is accesable now but, it's not drawing the charts when I press start.
ASSISTANT: It seems like the charts are not being updated because the `updateCharts` method is not being called when new data is logged. We need to modify the `logData` method in `AiderTest/AiderTest/ChartViewController.swift` to call `updateCharts` after appending new data to `weightData` and `flowData`.

Here are the changes:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    func logData(time: TimeInterval, weight: Float, flow: Float) {
        // Append the new data to the chart data
        weightData.append(ChartDataEntry(x: time, y: Double(weight)))
        flowData.append(ChartDataEntry(x: time, y: Double(flow)))

        // Update the charts
        updateCharts()
    }
=======
    func logData(time: TimeInterval, weight: Float, flow: Float) {
        // Append the new data to the chart data
        weightData.append(ChartDataEntry(x: time, y: Double(weight)))
        flowData.append(ChartDataEntry(x: time, y: Double(flow)))

        // Update the charts
        DispatchQueue.main.async {
            self.updateCharts()
        }
    }
>>>>>>> UPDATED
```

This change ensures that the charts are updated on the main thread, which is necessary because UI updates should always be done on the main thread.