aider: Updated the `toggleChartDisplay` function to switch between weight and flow charts in the `ChartViewController`.
# Context:
USER: I'm still seeing two seperate charts for our weight chart and flow chart. I want one chart that displays either the flow line or the weight line and be able to switch between them using our long tap gesture menu
ASSISTANT: To implement this, we need to modify the `toggleChartDisplay` function in `AiderTest/AiderTest/ViewController.swift` to switch between the weight and flow charts. However, the exact implementation will depend on how your `ChartViewController` is set up.

Assuming you have a `displayChart` function in `ChartViewController` that takes a parameter to determine which chart to display, here's how you could modify the `toggleChartDisplay` function:

```swift
AiderTest/AiderTest/ViewController.swift
<<<<<<< ORIGINAL
    func toggleChartDisplay() {
        isShowingWeightChart = !isShowingWeightChart
        // TODO: Update the chart view to display the correct chart based on the isShowingWeightChart variable
    }
=======
    func toggleChartDisplay() {
        isShowingWeightChart = !isShowingWeightChart
        if let chartViewController = navigationController?.viewControllers.first(where: { $0 is ChartViewController }) as? ChartViewController {
            chartViewController.displayChart(isShowingWeightChart ? .weight : .flow)
        }
    }
>>>>>>> UPDATED
```

In this code, `.weight` and `.flow` are assumed to be enumeration values that your `displayChart` function uses to determine which chart to display. You'll need to replace these with the actual values used in your `ChartViewController`.

Please note that this is a general approach and might need to be adjusted based on your actual implementation. If you provide the code for `ChartViewController`, I can give a more accurate solution.