aider: Modified 'displayChart' method to take an argument and implement chart displaying logic.
# Context:
USER: it seems you're not implimenting the logic to draw and display the chart. can you do that and the chart type should be a linechartview
ASSISTANT: To implement the logic to draw and display the chart, you need to modify the `displayChart` method in the `ChartViewController.swift` file.

Here's how you can modify the `displayChart` method:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    // Modify 'displayChart' method to take an argument
    func displayChart(_ type: ChartType) {
        // Implement your chart displaying logic here
    }
=======
    // Modify 'displayChart' method to take an argument and implement chart displaying logic
    func displayChart(_ type: ChartType) {
        switch type {
        case .weight:
            let weightDataSet = LineChartDataSet(entries: weightData, label: "Weight")
            weightDataSet.colors = [NSUIColor.blue]
            weightDataSet.drawCirclesEnabled = false // Disable dots
            weightDataSet.mode = .cubicBezier // Enable cubic bezier curve
            weightChart.data = LineChartData(dataSet: weightDataSet)
        case .flow:
            let flowDataSet = LineChartDataSet(entries: flowData, label: "Flow")
            flowDataSet.colors = [NSUIColor.red]
            flowDataSet.drawCirclesEnabled = false // Disable dots
            flowDataSet.mode = .cubicBezier // Enable cubic bezier curve
            flowChart.data = LineChartData(dataSet: flowDataSet)
        }
    }
>>>>>>> UPDATED
```

In this modification, I'm assuming that `ChartType` is an enum with two cases: `.weight` and `.flow`. If `ChartType` is defined differently, you need to adjust the switch statement accordingly.

Also, I'm using the `weightData` and `flowData` arrays to create the data sets for the charts. If these arrays are not supposed to be used for this purpose, you need to replace them with the appropriate data sources.

Finally, I'm setting the data of `weightChart` and `flowChart` to be the created data sets. If `weightChart` and `flowChart` are not supposed to display these data sets, you need to adjust the assignments accordingly.