aider: Modified 'displayChart' method to take 'weightData' or 'flowData' as an argument and implement chart displaying logic.
# Context:
USER: the `displayChart`
ASSISTANT: I see that you're referring to the `displayChart` method in the `ChartViewController.swift` file. However, I need more information to assist you better. Could you please provide more details about the changes you want to make to the `displayChart` method?
USER: the `displayChart` method should not take in chart type instead it should take in either the `weightData` or `flowData` because that will automatically cascade through to filling the dataset for our chart and therefor displaying the correct chart.
ASSISTANT: I understand. You want the `displayChart` method to take either `weightData` or `flowData` as an argument instead of `ChartType`. Here's how you can modify the `displayChart` method:

```swift
AiderTest/AiderTest/ChartViewController.swift
<<<<<<< ORIGINAL
    // Modify 'displayChart' method to take an argument and implement chart displaying logic
    func displayChart(_ type: ChartType) {
        switch type {
        case .weight:
            let weightDataSet = LineChartDataSet(entries: weightData, label: "Weight")
            weightDataSet.colors = [NSUIColor.blue]
            weightDataSet.drawCirclesEnabled = false // Disable dots
            weightDataSet.mode = .cubicBezier // Enable cubic bezier curve
            weightChart.data = LineChartData(dataSet: weightDataSet)
        case .flow:
            let flowDataSet = LineChartDataSet(entries: flowData, label: "Flow")
            flowDataSet.colors = [NSUIColor.red]
            flowDataSet.drawCirclesEnabled = false // Disable dots
            flowDataSet.mode = .cubicBezier // Enable cubic bezier curve
            flowChart.data = LineChartData(dataSet: flowDataSet)
        }
    }
=======
    // Modify 'displayChart' method to take 'weightData' or 'flowData' as an argument
    func displayChart(_ data: [ChartDataEntry], for chart: LineChartView, with label: String, color: NSUIColor) {
        let dataSet = LineChartDataSet(entries: data, label: label)
        dataSet.colors = [color]
        dataSet.drawCirclesEnabled = false // Disable dots
        dataSet.mode = .cubicBezier // Enable cubic bezier curve
        chart.data = LineChartData(dataSet: dataSet)
    }
>>>>>>> UPDATED
```

In this modification, I've changed the `displayChart` method to take an array of `ChartDataEntry` (either `weightData` or `flowData`), a `LineChartView` (either `weightChart` or `flowChart`), a label for the data set, and a color for the data set as arguments. You can call this method with either `weightData` and `weightChart` or `flowData` and `flowChart` to display the corresponding chart.